mongoDB에서 직접 조회 방식으로 채팅방 메타 정보(안읽은 수, 마지막 메시지) 반환하도록 변경 - 채팅방 나가는 시간 기준 조회

+++ src/main/java/smu/capstone/domain/chatroom/dto/ChatRoomEnterDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 public class ChatRoomEnterDto {
     private String userId;
     private RoomParticipantDto participant;
-    private int otherUserUnreadCount;
+    private Long otherUserUnreadCount;
 }
